TIP-1073: move ValueCollectionFactory and WriteValueCollectionFactory
