+++ src/Akeneo/Pim/Enrichment/Component/Product/Factory/Read/ValueFactory.php
@@ -10,6 +10,9 @@ use Akeneo\Pim\Structure\Component\Query\PublicApi\AttributeType\Attribute;
 use Webmozart\Assert\Assert;
 
 /**
+ * @TODO: move this class in Akeneo\Pim\Enrichment\Component\Product\Factory when the ValueFactory in this namespace
+ * will be deleted at the end of TIP-1073
+ *
  * @author    Anael Chardan <anael.chardan@akeneo.com>
  * @copyright 2019 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
