+++ tests/back/Pim/Enrichment/Specification/Bundle/EventSubscriber/EntityWithValues/LoadEntityWithValuesSubscriberSpec.php
@@ -5,7 +5,7 @@ namespace Specification\Akeneo\Pim\Enrichment\Bundle\EventSubscriber\EntityWithV
 use Doctrine\Common\Persistence\Event\LifecycleEventArgs;
 use PhpSpec\ObjectBehavior;
 use Akeneo\Pim\Enrichment\Bundle\EventSubscriber\EntityWithValues\LoadEntityWithValuesSubscriber;
-use Akeneo\Pim\Enrichment\Component\Product\Factory\Read\WriteValueCollectionFactory;
+use Akeneo\Pim\Enrichment\Component\Product\Factory\WriteValueCollectionFactory;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\WriteValueCollection;
